Code refactor. NavBar view with dropdown menu, styles and RWD finished. Also done general styles and static files adapted.
