+++ src/components/LoginInput/LoginInput.css
@@ -1,12 +1,12 @@
 input {
-    background-color: #292a2b;
-    width: 100%;
     height: 40px;
+    width: 100%;
     padding: 8px;
     margin: 0 auto 14px;
     border: 0;
-    outline: none;
+    background-color: #292a2b;
     align-items: flex-end;
+    outline: none;
     transition: background-color 0.5s;
 }
