+++ src/components/LoginInput/LoginInput.tsx
@@ -1,21 +1,23 @@
+import React from 'react';
 import './LoginInput.css';
 
 interface Props {
-  type: string;
-  placeholder: string;
-  value: string;
-  setValue(value: string): void;
+    type: string;
+    placeholder: string;
+    value: string;
+
+    setValue(value: string): void;
 }
 
 export const LoginInput = (props: Props) => {
-  return (
-    <input
-      type={props.type}
-      placeholder={props.placeholder}
-      value={props.value}
-      onChange={(e) => {
-        props.setValue(e.target.value);
-      }}
-    />
-  );
+    return (
+        <input
+            type={props.type}
+            placeholder={props.placeholder}
+            value={props.value}
+            onChange={(e) => {
+                props.setValue(e.target.value);
+            }}
+        />
+    );
 };
