--- src/components/PersonLabel/PersonLabel.css
@@ -1,38 +1,0 @@
-.person-label {
-    position: relative;
-    background-color: #1e1e1f;
-}
-
-.person-label--wrap, .person-label--wrap p, img, .filled-arrow {
-    background-color: inherit;
-}
-
-.person-label--wrap {
-    display: flex;
-    padding: 0 8px;
-    height: 50px;
-    align-items: center;
-    cursor: pointer;
-}
-
-.person-label--wrap img {
-    margin: 0 10px 0 0;
-    width: 40px;
-    border-radius: 50%;
-}
-
-.person-label--wrap p {
-    margin-right: 14px;
-}
-
-.person-label--wrap--visible {
-    background-color: rgb(37, 37, 37);
-}
-
-.person-label--wrap:hover {
-    background-color: rgb(47, 47, 47);
-}
-
-
-
-
