+++ src/components/UserMenu/UserMenu.css
@@ -1,4 +1,4 @@
-.person-menu {
+.user-menu {
     display: block;
     padding: 8px;
     position: absolute;
@@ -15,26 +15,27 @@
     transition: opacity 150ms ease-in-out, transform 150ms ease-in-out;
 }
 
-.person-menu--visible {
+.user-menu--visible {
     opacity: 1;
     transform: translateY(0);
     pointer-events: auto;
 }
 
-.person-menu a {
+.user-menu a {
     display: block;
-    background-color: rgb(40, 40, 40);
-    margin: 2px;
     padding: 8px;
+    margin: 2px;
     border-bottom: 1px solid #222224;
+    background-color: rgb(40, 40, 40);
     text-align: center;
     cursor: pointer;
+    transition: background-color 0.5s;
 }
 
-.person-menu a:hover {
+.user-menu a:hover {
     background-color: rgb(62, 62, 62);
 }
 
-.person-menu a:active {
-    background-color: rgb(70, 70, 70);
+.user-menu a:active {
+    background-color: rgb(90, 90, 90);
 }
