+++ src/components/common/FilledArrow/FilledArrow.css
@@ -4,6 +4,7 @@
     border-left: 4px solid transparent;
     border-right: 4px solid transparent;
     border-top: 7px solid #9e9e9e;
+    background-color: inherit;
     transition: transform 200ms ease-in-out;
 }
