--- src/components/common/MenuBar/MenuBar.tsx
@@ -1,12 +1,0 @@
-import {Logo} from '../Logo/Logo';
-import {PersonLabel} from '../../PersonLabel/PersonLabel';
-import './MenuBar.css';
-
-export const MenuBar = () => {
-    return (
-        <div className='menu-bar'>
-            <Logo width='70px' margin='auto 0'/>
-            <PersonLabel/>
-        </div>
-    );
-};
