+++ src/components/common/NavBar/NavBar.css
@@ -1,14 +1,14 @@
-.menu-bar {
+.nav-bar {
     display: flex;
     width: 100%;
-    justify-content: space-between;
     padding: 10px 14vw;
     background-color: #1e1e1f;
+    justify-content: space-between;
     position: fixed;
 }
 
 @media (max-width: 425px) {
-    .menu-bar {
+    .nav-bar {
         padding: 10px 4vw;
     }
 }
