+++ src/components/common/NavBar/NavBar.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import {Logo} from '../Logo/Logo';
+import {UserLabel} from '../../UserLabel/UserLabel';
+import './NavBar.css';
+
+export const NavBar = () => {
+    return (
+        <div className='nav-bar'>
+            <Logo width='70px' margin='auto 0'/>
+            <UserLabel/>
+        </div>
+    );
+};
