+++ src/components/common/RedButton/RedButton.css
@@ -1,17 +1,17 @@
 button {
-    background-color: #e02735;
-    border: none;
-    font-weight: 200;
     padding: 8px;
+    background-color: #e02735;
+    font-weight: 200;
+    border: none;
     transition: background-color 0.2s, transform 0.1s;
 }
 
 button:hover {
-    cursor: pointer;
     background-color: #cd0d1c;
+    cursor: pointer;
 }
 
 button:active {
-    transform: scale(.97)
+    transform: scale(.98)
 }
