+++ src/components/common/RedButton/RedButton.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import './RedButton.css';
 
 interface Props {
