+++ src/components/layouts/login/Login.tsx
@@ -1,12 +1,13 @@
+import React from 'react';
 import {LoginForm} from '../../LoginForm/LoginForm';
-import './Login.css';
 import {Logo} from '../../common/Logo/Logo';
+import './Login.css';
 
 export const Login = () => {
-  return (
-      <div className='login-wrap'>
-          <Logo/>
-          <LoginForm/>
-      </div>
-  );
+    return (
+        <div className='login-wrap'>
+            <Logo/>
+            <LoginForm/>
+        </div>
+    );
 };
